+++ lib/voicemail/call_controllers/mailbox_set_greeting_controller.rb
@@ -35,7 +35,7 @@ module Voicemail
 
     def record_greeting
       play config.set_greeting.before_record
-      record_comp = record config.set_greeting.recording.to_hash
+      record_comp = record record_options
       @temp_recording = record_comp.complete_event.recording.uri
       play_audio @temp_recording